Update 01. Area of Circle.java

+++ Problems-&-Solutions/Formula Related Problems/01. Area of Circle.java
@@ -7,7 +7,7 @@ class Main {
         Scanner scan = new Scanner(System.in);
         System.out.print("Enter radius of the circle :");
         int r = scan.nextInt();
-        double area = 3.14*r*r;
+        double area = (double)3.14*r*r;
         System.out.println("Area of a circle : "+area);
     }
 }
